finally got star rating to work, needed a new jquery.  Added Further examples of ads to go with products and cocktails.  Really want to make sure I am happy with the product setup. I have somemore work to do on that to get it looking decent as a data structure.

+++ src/model/cocktail.go
@@ -30,7 +30,7 @@ type Cocktail struct {
 	Tool            []Product
 	SourceName      string
 	SourceLink      string
-	Rating          float32
+	Rating          int
 	Flavor          []Meta
 	Type            []Meta
 	BaseSpirit      []Meta
